解决floor不适用于uint8型数据问题

+++ Decode.m
@@ -3,12 +3,12 @@
 function res = Decode( data , blocksize)
 
 [~,length] = size(data);
-i = 1;%the index of data
+i = uint16(1);%the index of data
 res = [];
 sum = 0;%store the sum of num in order to judge the end of the locate-map
-j = 1;%index of res
+j = uint16(1);%index of res
 while sum < ((blocksize-2)*(blocksize-2))/2 %locate-map
-   num = uint8(data(i)*64+data(i+1)*32+data(i+2)*16+data(i+3)*8+data(i+4)*4+data(i+5)*2+data(i+6));%calculate the number 
+   num = uint16(data(i)*64+data(i+1)*32+data(i+2)*16+data(i+3)*8+data(i+4)*4+data(i+5)*2+data(i+6));%calculate the number 
    sum = sum+num;
    res(j:j+num-1) = data(i+7);
    i = i+8;
@@ -16,7 +16,7 @@ while sum < ((blocksize-2)*(blocksize-2))/2 %locate-map
 end
 bit = [];
 sum = 0;
-p = length;
+p = uint16(length);
 q = 1;
 while sum < 4*blocksize-4 %MSB of the adjustment area 
     num = data(p-1)+data(p-2)*2+data(p-3)*4+data(p-4)*8+data(p-5)*16+data(p-6)*32+data(p-7)*64;
