+++ SaveSpace.m
@@ -32,7 +32,7 @@ for i = 1 : m
         for p = x+1 : 1 : x+blocksize-2
             for q = y+1 : 2 : y+blocksize-2
                 l = floor((embed_image(p,q)+embed_image(p,q+1))/2);
-                h = embed_image(p,q)-embed_image(p,q+1);
+                h = double(embed_image(p,q)-embed_image(p,q+1));
                 min = 2*(255-l);
                 b = 2*l+1;
                 if( b < min )
